ssh_server_port_int is unsigned short
It can never equal -1, and thus its not valid to compare it against -1.
This also means that this while loop would ignore the second part of the
test (since it would always return true).

Remove the if statement underneath that looks for -1, it will never
return true.

src/malware/CidSSH.cpp:99:62: warning: result of comparison of constant -1 with expression of type 'unsigned short' is always true [-Wtautological-constant-out-of-range-compare]
    while (*ssh_server_port_int == 0 && *ssh_server_port_int != -1) {
                                        ~~~~~~~~~~~~~~~~~~~~ ^  ~~
src/malware/CidSSH.cpp:103:30: warning: result of comparison of constant -1 with expression of type 'unsigned short' is always false [-Wtautological-constant-out-of-range-compare]
    if (*ssh_server_port_int == -1) { // could not bind any port for the SSH server
        ~~~~~~~~~~~~~~~~~~~~ ^  ~~