Add connection handler class
This creates the TCP/IP connection, and if necessary does this by
tunneling over a proxy. When the sockethandler callback is called the
socket is fully connected to the remote endpoint either through a proxy
or directly.

It is then ready to start talking whatever protocol is expected on the
other end of the line...